Delete old unused (and broken in some cases) functions from filter. Everything that was happening here is probably better to just get from scipy and/or sci-kit image.
